fix: 퀴즈 보기에 option_id 추가 -> 전체 기능 로직 수정

+++ backend/quiz/src/main/java/com/boindang/quiz/presentation/dto/response/WrongAnswerResponse.java
@@ -6,8 +6,8 @@ public record WrongAnswerResponse(
 	Long quizId,
 	String question,
 	List<String> options,
-	Long answerId,
-	Long selectedId,
+	int answerId,
+	int selectedId,
 	String explanation,
 	String selectedExplanation
 ) {}
